split common lib ast
